move bpf_loader helpers to tests
